requirements.txt
PyQt6==6.5.3
PyQt6-Qt6==6.5.3
websockets==12.0
volcengine-python-sdk[ark]
python-dotenv>=1.0.1
requests>=2.28.0